Refactor data repository location to support fake respository
